+++ DTTableViewManager/DTTableViewManager.swift
@@ -75,10 +75,6 @@ public class DTTableViewManager : NSObject {
         return TableViewFactory(tableView: self.tableView!)
     }()
     
-    deinit {
-        delegate = nil
-    }
-    
     /// Bundle to search your xib's in. This can sometimes be useful for unit-testing. Defaults to NSBundle.mainBundle()
     public var viewBundle = NSBundle.mainBundle()
     {
@@ -148,7 +144,7 @@ public class DTTableViewManager : NSObject {
     /// Call this method to retrieve model from specific UITableViewCell subclass.
     /// - Note: This method uses UITableView `indexPathForCell` method, that returns nil if cell is not visible. Therefore, if cell is not visible, this method will return nil as well.
     /// - SeeAlso: `StorageProtocol` method `objectForCell:atIndexPath:` - will return model even if cell is not visible
-    public func objectForVisibleCell<T:ModelTransfer where T:UITableViewCell>(cell:T?) -> T.ModelType?
+    public func itemForVisibleCell<T:ModelTransfer where T:UITableViewCell>(cell:T?) -> T.ModelType?
     {
         guard cell != nil else {  return nil }
         
@@ -163,7 +159,7 @@ public class DTTableViewManager : NSObject {
     /// - Parameter indexPath: NSIndexPath of the data model
     /// - Returns: data model that belongs to this index path.
     /// - Note: Method does not require cell to be visible, however it requires that storage really contains object of `ModelType` at specified index path, otherwise it will return nil.
-    public func objectForCellClass<T:ModelTransfer where T:UITableViewCell>(cellClass: T.Type, atIndexPath indexPath: NSIndexPath) -> T.ModelType?
+    public func itemForCellClass<T:ModelTransfer where T:UITableViewCell>(cellClass: T.Type, atIndexPath indexPath: NSIndexPath) -> T.ModelType?
     {
         return self.storage.itemForCellClass(T.self, atIndexPath: indexPath)
     }
@@ -173,7 +169,7 @@ public class DTTableViewManager : NSObject {
     /// - Parameter indexPath: NSIndexPath of the view
     /// - Returns: data model that belongs to this view
     /// - Note: Method does not require header to be visible, however it requires that storage really contains object of `ModelType` at specified section index, and storage to comply to `HeaderFooterStorageProtocol`, otherwise it will return nil.
-    public func objectForHeaderClass<T:ModelTransfer where T:UIView>(headerClass: T.Type, atSectionIndex sectionIndex: Int) -> T.ModelType?
+    public func itemForHeaderClass<T:ModelTransfer where T:UIView>(headerClass: T.Type, atSectionIndex sectionIndex: Int) -> T.ModelType?
     {
         return self.storage.itemForHeaderClass(T.self, atSectionIndex: sectionIndex)
     }
@@ -183,7 +179,7 @@ public class DTTableViewManager : NSObject {
     /// - Parameter indexPath: NSIndexPath of the view
     /// - Returns: data model that belongs to this view
     /// - Note: Method does not require footer to be visible, however it requires that storage really contains object of `ModelType` at specified section index, and storage to comply to `HeaderFooterStorageProtocol`, otherwise it will return nil.
-    public func objectForFooterClass<T:ModelTransfer where T:UIView>(footerClass: T.Type, atSectionIndex sectionIndex: Int) -> T.ModelType?
+    public func itemForFooterClass<T:ModelTransfer where T:UIView>(footerClass: T.Type, atSectionIndex sectionIndex: Int) -> T.ModelType?
     {
         return self.storage.itemForFooterClass(T.self, atSectionIndex: sectionIndex)
     }
@@ -694,5 +690,56 @@ extension DTTableViewManager : StorageUpdating
     private func controllerDidUpdateContent()
     {
         (self.delegate as? DTTableViewContentUpdatable)?.afterContentUpdate()
+    }
+}
+
+/// Deprecated methods
+extension DTTableViewManager
+{
+    /// Call this method to retrieve model from specific UITableViewCell subclass.
+    /// - Note: This method uses UITableView `indexPathForCell` method, that returns nil if cell is not visible. Therefore, if cell is not visible, this method will return nil as well.
+    /// - SeeAlso: `StorageProtocol` method `objectForCell:atIndexPath:` - will return model even if cell is not visible
+    @available(*, unavailable, renamed="itemForVisibleCell")
+    public func objectForVisibleCell<T:ModelTransfer where T:UITableViewCell>(cell:T?) -> T.ModelType?
+    {
+        guard cell != nil else {  return nil }
+        
+        if let indexPath = tableView?.indexPathForCell(cell!) {
+            return storage.itemAtIndexPath(indexPath) as? T.ModelType
+        }
+        return nil
+    }
+    
+    /// Retrieve model of specific type at index path.
+    /// - Parameter cell: UITableViewCell type
+    /// - Parameter indexPath: NSIndexPath of the data model
+    /// - Returns: data model that belongs to this index path.
+    /// - Note: Method does not require cell to be visible, however it requires that storage really contains object of `ModelType` at specified index path, otherwise it will return nil.
+    @available(*, unavailable, renamed="itemForCellClass")
+    public func objectForCellClass<T:ModelTransfer where T:UITableViewCell>(cellClass: T.Type, atIndexPath indexPath: NSIndexPath) -> T.ModelType?
+    {
+        return self.storage.itemForCellClass(T.self, atIndexPath: indexPath)
+    }
+    
+    /// Retrieve model of specific type for section index.
+    /// - Parameter headerView: UIView type
+    /// - Parameter indexPath: NSIndexPath of the view
+    /// - Returns: data model that belongs to this view
+    /// - Note: Method does not require header to be visible, however it requires that storage really contains object of `ModelType` at specified section index, and storage to comply to `HeaderFooterStorageProtocol`, otherwise it will return nil.
+    @available(*, unavailable, renamed="itemForHeaderClass")
+    public func objectForHeaderClass<T:ModelTransfer where T:UIView>(headerClass: T.Type, atSectionIndex sectionIndex: Int) -> T.ModelType?
+    {
+        return self.storage.itemForHeaderClass(T.self, atSectionIndex: sectionIndex)
+    }
+    
+    /// Retrieve model of specific type for section index.
+    /// - Parameter footerView: UIView type
+    /// - Parameter indexPath: NSIndexPath of the view
+    /// - Returns: data model that belongs to this view
+    /// - Note: Method does not require footer to be visible, however it requires that storage really contains object of `ModelType` at specified section index, and storage to comply to `HeaderFooterStorageProtocol`, otherwise it will return nil.
+    @available(*, unavailable, renamed="itemForFooterClass")
+    public func objectForFooterClass<T:ModelTransfer where T:UIView>(footerClass: T.Type, atSectionIndex sectionIndex: Int) -> T.ModelType?
+    {
+        return self.storage.itemForFooterClass(T.self, atSectionIndex: sectionIndex)
     }
 }